+++ parler/utils/conf.py
@@ -17,7 +17,7 @@ class LanguagesSetting(dict):
         For an example, see :func:`~parler.appsettings.add_default_language_settings`.
         """
         if site_id is None:
-            site_id = settings.SITE_ID
+            site_id = getattr(settings, 'SITE_ID', None)
 
         for lang_dict in self.get(site_id, ()):
             if lang_dict['code'] == language_code: